backend/requirements.txt
Django>=4.2,<5.0
huggingface-hub>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
